change: Move member.h from syntax/ to member/
