+++ include/absent/member/member.h
@@ -1,7 +1,7 @@
 #ifndef RVARAGO_ABSENT_MEMBER_H
 #define RVARAGO_ABSENT_MEMBER_H
 
-namespace rvarago::absent::syntax::member {
+namespace rvarago::absent::member {
     template<typename A, typename B>
     using Mapper = B (A::*)() const;
 }
